% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\docType{data}
\name{beeton}
\alias{beeton}
\title{Recipes from Mrs. Beeton's Dictionary of Every-Day Cookery}
\format{
A tibble with 24102 rows of character strings
}
\source{
\url{https://www.gutenberg.org/cache/epub/50191/pg50191.txt}
}
\usage{
beeton
}
\description{
A dataset containing the text of every recipe in "Mrs. Beeton's Dictionary
of Every-Day Cookery". The UTF-8 plain text was sourced from Project
Gutenberg. Any text elements not associated with a recipe was removed.
}
\keyword{datasets}
