% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_verses.R
\name{get_verses}
\alias{get_verses}
\title{Get a random string with exactly 17 syllables}
\usage{
get_verses(book)
}
\arguments{
\item{book}{Input text. A dataframe or tibble, with character strings in
each row}
}
\value{
A character string with 17 syllables
}
\description{
Get a random string with exactly 17 syllables
}
